+++ chat-server/chat-server.cpp
@@ -1,47 +1,8 @@
 // chat-server.cpp : This file contains the 'main' function. Program execution begins and ends there.
 //
+#include "defs.hpp"
 
-#include <assert.h>
-
-// Copied for work we've done before
-#if defined(_WIN32) || defined(_WIN64)
-// Windows
-#define WIN32_LEAN_AND_MEAN 1
-#define NOMINMAX 1
-#define  _WINSOCK_DEPRECATED_NO_WARNINGS 1
-#include <WinSock2.h>
-#include <ws2tcpip.h>
-#pragma comment (lib, "Mswsock.lib")
-#pragma comment (lib, "Ws2_32.lib")
-#pragma comment (lib, "AdvApi32.lib")
-
-class WsaWrapper {
-public:
-	WsaWrapper() {
-		WSADATA data;
-		auto rc = WSAStartup(MAKEWORD(2, 2), &data);
-		assert(rc == 0);
-	}
-	~WsaWrapper() {
-		WSACleanup();
-	}
-};
-using socket_t = SOCKET;
-using ssize_t = int;
-
-#else
-// Unix
-#include <unistd.h>
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <netinet/ip.h>
-#include <arpa/inet.h>
-
-using socket_t = int;
-#endif
-
-using port_t = short;
-
+#include <cstring>
 #include <iostream>
 #include <map>
 #include <stdexcept>
@@ -62,7 +23,9 @@ void recvClientHello(socket_t s);
 // a very basic server
 int main()
 try {
+#if defined(_WIN32) || defined(_WIN64)
 	WsaWrapper wsa;
+#endif
 
 	int rc;
 	port_t port = 3232; // some random port
@@ -87,8 +50,13 @@ try {
 	// examine what was sent
 	nbytes = send(cli, "jan ok", 6, 0); assert(nbytes == 6);
 
+#if defined(_WIN32) || defined(_WIN64)
 	closesocket(cli);
 	closesocket(s);
+#else
+	close(cli);
+	close(s);
+#endif
 }
 catch (const std::exception& e) {
 	std::cerr << "fatal: " << e.what() << "\n";
